refactor: util 패키지 생성한 후 파일 위치 변경
