+++ src/main/java/lotto/util/NumberGenerator.java
@@ -1,6 +1,7 @@
-package lotto.domain;
+package lotto.util;
 
 import camp.nextstep.edu.missionutils.Randoms;
+import lotto.domain.LottoConfiguration;
 
 import java.util.ArrayList;
 import java.util.Collections;
